[Refactor] Split heavy config machinery out of configbase.h
* Add common/configbase_impl.h to host Field/Enum/Alias definitions,
  strtox helpers, and the defining versions of the CONF_* macros so
  configbase.h stays lightweight for most includes.
* remove the __IN_CONFIGBASE_CPP__ backdoor implementation
